Upgraded precision for Node Values

+++ Unity Project/Assets/Scripts/ECS/DataComponents/Tags/NodeData.cs
@@ -5,5 +5,5 @@ using Unity.Collections;
 public struct NodeData : IComponentData
 {
     public FixedString32 nodeName;
-    public float nodeValue;
+    public double nodeValue;
 }
